pyproject.toml
[tool.poetry]
name = "habits"
version = "0.1.0"
description = ""
authors = ["KarelinIvan <ivan.karelin.1993@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
eventlet = "^0.38.1"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
pytz = "^2024.2"
requests = "^2.32.3"
coverage = "^7.6.9"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
